+++ CHANGELOG.rst
@@ -10,6 +10,7 @@ This document describes changes between each past release.
 
 - List of recipients can now contain groups URIs. The principals from the specified
   group that look like email addresses will be used as recipients (fixes #6)
+- Support new variables like server root url or client IP address in email template (fixes #22)
 
 
 0.1.0 (2017-01-25)